Added basic article saving
As can be seen in test/models/article_test.rb, I added some basic
validation to ensure that all of the fields for Article are filled in.
The controller has been updated to save created records to the
database. The view has been updated to properly redirect, run “rake
routes” to see what articles_path does.